requirements.txt
click==8.0.4
dataclasses==0.8
Flask==2.0.3
importlib-metadata==4.8.3
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
paho-mqtt==1.6.1
pkg_resources==0.0.0
python-decouple==3.6
typing_extensions==4.1.1
waitress==2.0.0
Werkzeug==2.0.3
zipp==3.6.0